process-receipts/requirements.txt
functions-framework==3.4.0
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0